Audit teamkill events in rpt
Based on https://github.com/NeoArmageddon/co10_Escape/issues/133 when a
player teamkills another player, the script will emit a message in the
rpt.

This message can be used to investigate intentional teamkill events of
trolls or alike to verify claims of other players or to identify massive
or continued teamkill events.